+++ Changelog.md
@@ -2,6 +2,15 @@
 
 ## [Not Released v0.9]
 
+## 2022-01-31/ml
+- Pagination Bar is shorter 
+- Login/Logout Button in ng-container to delete out of dom-structure
+- visible var in static service
+
+## 2022-01-27/avz
+### added
+- Textcomponent with video on the teacher info page
+
 ## 2022-01-27/avz
 ### changed
 - added navigation bar to the header component